requirements.txt
onnxruntime-gpu
onnx
pillow
numpy
opencv-python
boto3==1.26.142
boto3-stubs[essential]==1.26.142
python-dotenv
mariadb